Enable Android edge-to-edge rendering and consistently show status bar
This updates the app to render underneath the status and nav bars on Android, following the [best practice guidelines]() (and looking nicer). This also ensures the status bar is always visible on Android to no longer hide folks' notifications. I've tested this on my phone by navigating the main lists, viewing tags (PDF and image based), videos for tags, using the app drawer, searching for tags; and doing that all in both portrait and landscape orientation.

I've put some images in the PR to demonstrate the difference.

Breaking down the changes a bit more, in the `android` directory this PR:
- Fixing the Android build to not try to get the release signing configs each time
- Allowing side-by-side installs of the debug and release versions of the app on Android (`applicationIdSuffix`)
- Using the `edgeToEdge` function from `androidx.activity` library to setup most of the edge-to-edge functionality (this library required bumping the compile SDK version but that's quite a safe thing to change).
- Unifying the themes and adding a few extra bits for edge-to-edge (telling the OS the status bar will typically be a light color (so it can render dark icons) and letting us layout underneath the display cutouts in landscape mode).

And in the `src` directory this PR:
- Adjusts the list header to use the insets for height in portrait and provide a buffer in landscape ("shallow")
- Add padding to the sheetmusic view so it initially renders without going under the system bars/"unsafe" areas.
- Add padding/margins to all the other UI elements on the tag page (back button, favorite banner, FAB, bottom controls, video modal close button)
- Add padding to the tab bar to avoid the left/right "unsafe" areas and the bottom nav bar.
- Don't hide the status bar on Android (it's annoying!)

Some areas of further improvement, either in this PR or a future one:
- There's a few places where do actually end up drawing a dark background for the status bar, rather than a light one, but because we've told the OS we're using a light color it uses black icons which get hard to see. Ideally in these screens (namely the tag screen modals and the welcome screen) we would dynamically tell the OS we're using a dark color.
- It would be awesome to support dark mode, which would tie into the above point about light/dark status bar, though honestly wouldn't have much more relevance to this PR.

I'm definitely happy to make any changes you'd like to see here or answer any questions about what I've done!