[Offline search] Add a script to periodically fetch offline DB, host on GitHub Pages
This is the first phase in modifying goodtags for instant offline search. Specifically this handles the "server side" issue of regularly generating reasonably up-to-date database snapshots that the app can download on demand. Subsequent PRs will have changes in the app itself to actually download that database and switch out the part that currently hits the API to instead perform a query on the local SQL database.

This adds a Python script which will regularly (currently once a week) pull down the entirety of the Barbershop Tags database and put it into a SQL database suitable for consumption by the app (as mentioned, coming in a subsequent PR). I think by doing this in batches, weekly, and in the middle of the night (in the US) we won't put too high strain on the database, though that's making a number of assumptions like that their peak load tends to be during the day in the US. Note that I'd considered only trying to fetch what's new since the last fetch, but since download count increments mark a row as "new", it ends up being the case that after a week almost half of the rows are new anyway so the added complexity of trying to do the fetch incrementally didn't seem worth it.

The script then pushes the SQL database and a manifest (which can be used by the app to determine if the remote database is actually newer) to a `gh-pages` branch, which will then be used to host a GitHub Pages, well, page. GH Pages is used instead of `raw.githubusercontent.com` because the former supports `Accept-Encoding: gzip` which should transparently and meaningfully cut down on transfer sizes of the DB since it compresses down pretty well.

Additionally this creates a GitHub Actions workflow which actually ends up triggering this script once a week.