[Offline search] Ensure the up-to-date database is available in the app
This is the second phase in modifying goodtags for ~instant offline search. To recap, at a high level, the cumulative set of changes will look like:
1. Have something that periodically snapshots the canonical barbershoptags.com database into a SQL database suitable for use by the app and host that where the app can download it (https://github.com/kamatsuoka/goodtags/pull/7).
2. Update the app to include the latest (at the time) database when compiling it and fetch the latest database on startup, only downloading it if it's actually newer than what the app already has (**this PR**).
3. Update the app to add a parallel code path to direct searches to the database rather than the API.
4. Once this has been validated, rip out the old code path and solely rely on the local database for searches.

There are two main changes in the PR: ensuring the latest database is downloaded at build time, and downloading the latest database at app startup time.

At build time there's a new `yarn download-search-db` task which pulls the latest DB into the `src/assets/generated_db/` folder. This task is then invoked as a prerequisite for a number of the other `yarn` tasks as well as in the gradle and xcode build configs. This also modifies the Metro config to include the `.sqlite` file as an asset in the bundle.

At app startup time, there's a new utility for transparently ensuring the database is available and kicking off a background update. This code path will copy the database from the app bundle produced by Metro into the app's local storage if there isn't already a database or the one that's already there is older (as determined by the accompanying manifest). It'll then read the remote manifest and, if that's newer, download that database and seamlessly switch out the backing database instance with the new file. That switching out of the existing database is facilitated by the new `DbWrapper` class which tracks live transactions and replaces the underlying database once there are no more transactions; it's effectively a RW lock on the database. All this code relies on promises (which return ~instantly once they're resolved) so the calling code doesn't need to know about any of those details, it just says "give me a database and run this transaction on it".

To test this, I've written some Jest tests for the `DbWrapper` class and have manually tested the in-app update flows:
- No existing local DB -> copy from bundle
- Bundle/app DB newer than local DB -> copy from bundle
- Bundle/app DB older than/same as local DB -> no copy
- Remote DB newer than local DB -> copy from remote
- Remote DB older than/same as local DB -> no copy

Ideally I'd automate these tests with detox but that setup seems to have atrophied and I wasn't able to get it working in a somewhat short amount of time. I may still look into fixing that up (and getting it running in CI) and then automating these tests, but that would be in a subsequent PR.