[Offline search] Hook up database for full offline search
This is the third phase in modifying goodtags for ~instant offline search. To recap, at a high level, the cumulative set of changes will look like:
1. Have something that periodically snapshots the canonical barbershoptags.com database into a SQL database suitable for use by the app and host that where the app can download it (https://github.com/kamatsuoka/goodtags/pull/7).
2. Update the app to include the latest (at the time) database when compiling it and fetch the latest database on startup, only downloading it if it's actually newer than what the app already has (https://github.com/kamatsuoka/goodtags/pull/9).
3. Update the app to add a parallel code path to direct searches to the database rather than the API (**this PR**).
4. Once this has been validated, rip out the old code path and solely rely on the local database for searches.

This PR adds a new box in the search dialog which allows selection of "mode" as either "offline" (default) or "online" and includes a button for a dialog that describes the difference a bit more. That setting feeds into the actual fetch utilities which will either take the existing code path if "online" is selected, or a brand new code path that hits the local SQLite database if "offline" is selected.

Since we're adding a new field to a persisted store that we want to default to a non-`undefined` value, this adds a migrations system for the persisted store following `redux-persist`'s guide at https://github.com/rt2zz/redux-persist/blob/master/docs/migrations.md. Note that the `InitialFilters` value is only relevant when there isn't any existing persisted store.

This makes a few changes to the internal search interface to unify how we interact with the API and the local DB. The search parameters are no longer given as the API expects them but rather using something that more closely resembles their semantic meaning; the translation to the API's query parameters is pushed one layer deeper. Similarly, the app now works on a 0-indexed system rather than the API's 1-indexed results. There are parallel functions for translating the database query results into tag objects, and while there's some duplication currently, the idea is to eventually remove the API/xml ones, so I didn't think it was worth trying to unify them.

Behavior-wise, the DB results are quite similar to the API's, though there are a few known changes:
- It's only searching for whole tokens, so searching for "ebb" will give "Ebb Tide" as expected, but not "Moonlight in Vermont" (which is contributed by "W**ebb** comfort")
- Since the tag ID is included in the set of tokens being searched, we don't do the ID special casing when searching locally, so looking for "1776" now turns up both the tag #1776 ("I Get The Blues Most Every Night") *and* the actual "1776" tag (#922).

This has been mostly tested manually. I've compared the results between this version of the app and the version that's currently live on the Play Store and have been using this offline version of the app for a few weeks. However I haven't written any automated tests since I'm not sure if I can get `expo-sqlite` working within Jest and Detox isn't currently an option.