Fix eslint errors and check eslint in CI
This runs `yarn run eslint --fix src` and commits the result. It also extends the CI test workflow to check that there are no eslint errors. Note I needed to bump the `typescript-eslint` packages because they weren't new enough to know about the version of TypeScript we use and so were printing warnings on every run.