Only download databases with valid schema versions
This extends the manifest format to have a mapping from schema version to db name and updates the app to only download databases that match the schema version it's expecting. This would allow us to publish multiple schema versions simultaneously if needed and helps ensure that bumping the schema in the future won't break existing apps (they'll just get more and more out of date once we stop generating that schema version).

This means that the download logic (both the `downloadLatestSearchDb.ts` and `sqlUtil.ts`) no longer need to know what the database is named on the remote, so the hacky `.otf` information is solely contained in the generation script. It also updates the two downloaders to share constants (most notably what schema version is valid) to reduce the chance of drift.